requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
httpx>=0.27.0
itsdangerous>=2.1.0
jinja2>=3.1.0
tzdata>=2024.1
psutil>=5.9.0
vaderSentiment>=3.3.0
wavelink>=3.4.0
spotipy>=2.23.0
pynacl>=1.5.0
